requirements.txt
Django==1.8.1
Mezzanine==4.1.0
Pillow==2.8.1
PyJWT==1.4.0
Werkzeug==0.11.10
Whoosh==2.6.0
amqp==1.4.6
anyjson==0.3.3
beautifulsoup4==4.1.3
billiard==3.3.0.20
bleach==1.4.1
celery==3.1.23
chardet==2.3.0
dj-database-url==0.3.0
django-celery==3.1.17
django-contrib-comments==1.7.1
django-extensions==1.6.7
django-forms-bootstrap==3.0.1
django-haystack==2.5.dev1
django-jsonfield==0.9.13
django-paypal==0.2.5
django-social-auth==0.7.28
django-stripe-payments==2.0.0
filebrowser-safe==0.3.8
future==0.9.0
gnureadline==6.3.3
grappelli-safe==0.3.13
gunicorn==19.1.1
html5lib==0.99999
httplib2==0.9.1
kombu==3.0.26
lxml==3.4.0
oauth2==1.5.211
oauthlib==0.7.2
payments==0.02
paypalrestsdk==1.11.5
psycopg2==2.6
pyOpenSSL==16.0.0
pytesseract==0.1.5
python-openid==2.2.5
python-social-auth==0.2.19
pytz==2015.4
requests==2.10.0
requests-oauthlib==0.5.0
simplejson==3.8.2
six==1.9.0
stripe==1.35.0
tzlocal==1.0
wsgiref==0.1.2
django-colorfield==0.1.10